Merge/#12-싱글게임 구현

See merge request s10-bigdata-dist-sub2/S10P22A207!29

+++ backend/src/main/java/com/backend/api/domain/member/dto/response/ProfileSingleGameLogRes.java
@@ -14,7 +14,7 @@ public record ProfileSingleGameLogRes(
 	Double finalRoi,
 
 	@Schema(description = "최종 수익")
-	Integer finalProfit,
+	Long finalProfit,
 	@Schema(description = "게임 플레이 날짜")
 	String gameDate
 ) {
